move getKeyedListEntry() to util.go
